// src/proto/user/AdminRegisterResponse.ts
// Original file: src/proto/user.proto


export interface AdminRegisterResponse {
  'message'?: (string);
  'email'?: (string);
  'password'?: (string);
  'adminId'?: (string);
}

export interface AdminRegisterResponse__Output {
  'message'?: (string);
  'email'?: (string);
  'password'?: (string);
  'adminId'?: (string);
}